minify the main header file it reach to 300 line which was very hard to read and maintain and has several unrelated things in one place now it is cleaner shorter with 130 line and we wish more cleaner code
