+++ client/index.html
@@ -6,6 +6,7 @@
     <link
       href="https://fonts.googleapis.com/css2?family=Almarai:wght@300;400;700&display=swap"
       rel="stylesheet"
+      loading="lazy"
     />
 
     <link rel="icon" type="image/svg+xml" href="src/assets/mainLogo.svg" />
