+++ client/src/components/RootLayoutComponents/MainHeader.jsx
@@ -1,51 +1,26 @@
-import { useState, useEffect } from "react";
-import { Link, NavLink, useNavigate } from "react-router-dom";
+import { useState } from "react";
+import { Link, useNavigate } from "react-router-dom";
 import { Menu, X, Search } from "lucide-react";
 import main_logo from "../../assets/mainLogo.svg";
-import { useCourseData } from "../../context/CourseContext";
 import { useAuth } from "../../context/authContext";
 import { toast } from "react-toastify";
 import { motion } from "framer-motion";
 import ThreeDotMenu from "./ThreeDotMenu";
+import NavigationLink from "./NavigationLink";
+import SearchBar from "./SearchBarForMobile";
+import SearchBarForDesktop from "./SearchBarForDesktop";
 
 export default function MainHeader() {
   const navigate = useNavigate();
   const { logout, isAuthorized } = useAuth();
-  const [searchInput, setSearchInput] = useState("");
+
   const [isOpen, setIsOpen] = useState(false);
-  const [filterdCourses, setfilterdCourses] = useState(null);
-  const [showResults, setShowResults] = useState(false);
+
   const [isSearching, setIsSearching] = useState(false);
 
   function handleSearching() {
     setIsSearching(!isSearching);
   }
-
-  // Use the context to get courses data
-  const { coursesData, fetchCoursesContext } = useCourseData();
-
-  // Only fetch the data one time then use
-  useEffect(() => {
-    if (!coursesData) {
-      fetchCoursesContext();
-    }
-  }, [coursesData]);
-
-  const onSearch = (e) => {
-    const { value } = e.target;
-    setSearchInput(value);
-    setShowResults(value.length > 0);
-
-    if (coursesData) {
-      setfilterdCourses(
-        coursesData.filter(
-          (course) =>
-            course.code.toLowerCase().includes(value.toLowerCase()) ||
-            course.name.toLowerCase().includes(value.toLowerCase())
-        )
-      );
-    }
-  };
 
   const handleLogout = async () => {
     try {
@@ -94,127 +69,19 @@ export default function MainHeader() {
               isOpen ? "flex flex-col gap-4 mb-2" : "hidden"
             } md:flex-row md:justify-center`}
           >
-            <NavLink
-              className="text-gray-700 hover:text-gray-500 transition-colors w-full md:w-auto text-center "
-              to="/home"
-              end
-            >
-              {({ isActive }) => (
-                <div
-                  className={`relative w-full whitespace-nowrap md:w-auto text-center py-1 px-3 
-                    border-b-4 border-TAF-100 transition-all duration-300 
-                   ${isActive ? "border-opacity-100" : "border-opacity-0"}`}
-                >
-                  الصفحة الرئيسية
-                </div>
-              )}
-            </NavLink>
-            <NavLink
-              className="text-gray-700 hover:text-gray-500 transition-colors w-full md:w-auto text-center "
-              to="/courses"
-              end
-            >
-              {({ isActive }) => (
-                <div
-                  className={`relative w-full md:w-auto text-center py-1 px-3 
-                    border-b-4 border-TAF-100 transition-all duration-300 
-                   ${isActive ? "border-opacity-100" : "border-opacity-0"}`}
-                >
-                  المواد
-                </div>
-              )}
-            </NavLink>
-            <NavLink
-              className="text-gray-700 hover:text-gray-500 transition-colors w-full md:w-auto text-center "
-              to="/myquizzes"
-              end
-            >
-              {({ isActive }) => (
-                <div
-                  className={`relative w-full whitespace-nowrap md:w-auto text-center py-1 px-3 
-                    border-b-4 border-TAF-100 transition-all duration-300 
-                   ${isActive ? "border-opacity-100" : "border-opacity-0"}`}
-                >
-                  إختباراتي القصيرة
-                </div>
-              )}
-            </NavLink>
-            <NavLink
-              className="text-gray-700 hover:text-gray-500 transition-colors w-full md:w-auto text-center "
-              to="/mypreviousschedules"
-              end
-            >
-              {({ isActive }) => (
-                <div
-                  className={`relative w-full whitespace-nowrap md:w-auto text-center py-1 px-3 
-                    border-b-4 border-TAF-100 transition-all duration-300 
-                   ${isActive ? "border-opacity-100" : "border-opacity-0"}`}
-                >
-                  جداولي السابقة
-                </div>
-              )}
-            </NavLink>
+            <NavigationLink
+              linkTo={"الصفحة الرئيسية"}
+              route={isAuthorized ? "/home" : "/"}
+            />
+            <NavigationLink linkTo={"المواد"} route={"/courses"} />
+            <NavigationLink linkTo={"إختباراتي القصيرة"} route="myquizzes" />
+            <NavigationLink
+              linkTo={"جداولي السابقة"}
+              route={"mypreviousschedules"}
+            />
           </div>
         )}
-
-        {isAuthorized && (
-          <div
-            className={`w-full md:hidden ${
-              isOpen ? "flex flex-col gap-4 mb-4" : "hidden"
-            }`}
-          >
-            <div className="w-full relative">
-              <div className="relative flex items-center">
-                <Search className="absolute left-3 h-5 w-5 text-gray-400" />
-                <input
-                  type="text"
-                  placeholder="إبحث عن المواد"
-                  className="w-full p-2 pl-10 rounded-md border border-gray-300 focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-transparent transition-all"
-                  value={searchInput}
-                  onChange={onSearch}
-                  onFocus={() => setShowResults(searchInput.length > 0)}
-                  onBlur={() => setTimeout(() => setShowResults(false), 200)}
-                />
-              </div>
-
-              {/* Search Results */}
-              {showResults && (
-                <div className="absolute z-50 w-full mt-1 bg-white rounded-md border border-gray-300 shadow-lg">
-                  <div className="max-h-60 overflow-y-auto">
-                    {filterdCourses?.length === 0 ? (
-                      <div className="p-3 text-gray-500">لا توجد نتائج</div>
-                    ) : (
-                      <ul className="py-2">
-                        {filterdCourses?.map((course) => (
-                          <li
-                            key={course.id}
-                            className="px-3 py-2 hover:bg-gray-100 group/searchResult"
-                          >
-                            <Link
-                              onClick={() => {
-                                setIsSearching(false);
-                                setSearchInput("");
-                              }}
-                              to={`/courses/${course.id}`}
-                              className="block text-lg"
-                            >
-                              <p className="text-gray-700 group-hover/searchResult:text-blue-500">
-                                {course.name} |{" "}
-                                <span className="font-bold text-gray-900">
-                                  {course.code}
-                                </span>
-                              </p>
-                            </Link>
-                          </li>
-                        ))}
-                      </ul>
-                    )}
-                  </div>
-                </div>
-              )}
-            </div>
-          </div>
-        )}
+        {isAuthorized && <SearchBar isOpen={isOpen} />}
 
         {/* Sign Up Button mobile */}
         <div
@@ -252,70 +119,7 @@ export default function MainHeader() {
       </nav>
 
       {isAuthorized && isSearching && (
-        <motion.div
-          initial={{ opacity: 0, y: -100 }} // Start above the navigation bar
-          animate={{ opacity: 1, y: 0 }} // Slide down to its position
-          exit={{ opacity: 0, y: -100 }} // Slide back up when exiting
-          transition={{ duration: 0.2, ease: "easeOut" }}
-          className="hidden md:flex md:justify-center md:items-center p-2 border-b border-b-TAF-100 bg-TAF-200 z-50 
-               fixed top-[100px] left-0 w-full shadow-md" // Position below the navigation bar
-        >
-          <div className="w-3/5 relative">
-            <div className="relative flex items-center">
-              <Search className="absolute left-3 h-5 w-5 text-gray-400" />
-              <input
-                type="text"
-                placeholder="إبحث عن المواد"
-                className="w-full p-2 pl-10 rounded-md border border-gray-300 focus:outline-none 
-                     focus:ring-2 focus:ring-blue-500 focus:border-transparent transition-all"
-                value={searchInput}
-                onChange={onSearch}
-                onFocus={() => setShowResults(searchInput.length > 0)}
-                onBlur={() => setTimeout(() => setShowResults(false), 200)}
-              />
-            </div>
-            <div className="relative">
-              <button
-                onClick={handleSearching}
-                className="hidden md:absolute md:top-1 md:right-1"
-              >
-                <X size={18} />
-              </button>
-            </div>
-            {/* Search Results */}
-            {showResults && (
-              <div className="absolute z-50 w-full mt-1 bg-white rounded-md border border-gray-300 shadow-lg">
-                <div className="max-h-60 overflow-y-auto">
-                  {filterdCourses?.length === 0 ? (
-                    <div className="p-3 text-gray-500">لا توجد نتائج</div>
-                  ) : (
-                    <ul className="py-2">
-                      {filterdCourses?.map((course) => (
-                        <li
-                          key={course.id}
-                          className="px-3 py-2 hover:bg-gray-100 group/searchResult"
-                        >
-                          <Link
-                            onClick={() => setSearchInput("")}
-                            to={`/courses/${course.id}`}
-                            className="block text-lg"
-                          >
-                            <p className="text-gray-700 group-hover/searchResult:text-blue-500">
-                              {course.name} |{" "}
-                              <span className="font-bold text-gray-900">
-                                {course.code}
-                              </span>
-                            </p>
-                          </Link>
-                        </li>
-                      ))}
-                    </ul>
-                  )}
-                </div>
-              </div>
-            )}
-          </div>
-        </motion.div>
+        <SearchBarForDesktop handleSearching={handleSearching} />
       )}
     </div>
   );
