+++ client/tailwind.config.js
@@ -1,4 +1,6 @@
 export default {
+  mode: "jit",
+  content: ["./src/**/*.{js,jsx,ts,tsx}", "./public/index.html"],
   darkMode: ["class"],
   content: ["./index.html", "./src/**/*.{js,ts,jsx,tsx}"],
   theme: {
